Move SentenceFaker into `faker_extensions` package Make single module to hold all sentence semantic dependency logic for Faker, including SentenceFaker, RecordGenerator and RecordsFaker
