+++ presidio_evaluator/data_generator/__init__.py
@@ -7,7 +7,7 @@ if not hasattr(_raw_data_path, '__getitem__'):
     _raw_data_path = _raw_data_path._path
 raw_data_dir = Path(_raw_data_path[0])
 
-from .presidio_data_generator import SentenceFaker, PresidioSentenceFaker
+from .presidio_data_generator import PresidioSentenceFaker
 from .presidio_pseudonymize import PresidioPseudonymization
 
 
@@ -17,5 +17,5 @@ def read_synth_dataset():
     )
 
 
-__all__ = ["SentenceFaker", "PresidioSentenceFaker", "PresidioPseudonymization", "read_synth_dataset",
+__all__ = ["PresidioSentenceFaker", "PresidioPseudonymization", "read_synth_dataset",
            "raw_data_dir"]
